Start hidden option and icon update
Now comes with a "start hidden" option so if you use the program and do
not want to constantly hide the window, you can do so!

Added the icon from online:
http://www.iconfinder.com/iconsets/superpack#readme
Says to link to the author so I linked to:
http://www.iconfinder.com/iconsets/superpack#readme

Check out the updated readme for everything!

Changed layout of window so I can fit more options in... and put in the
"Made by: Rebant" and Icon credit. :)